Do not use direction package on macos
