+++ store/disk.go
@@ -1,8 +1,6 @@
 package store
 
 import (
-	"bytes"
-	"io"
 	"io/ioutil"
 	"os"
 	"path"
@@ -13,8 +11,6 @@ import (
 
 	"github.com/lbryio/lbry.go/v2/extras/errors"
 	"github.com/lbryio/lbry.go/v2/stream"
-
-	"github.com/brk0v/directio"
 )
 
 // DiskStore stores blobs on a local disk
@@ -74,40 +70,6 @@ func (d *DiskStore) Get(hash string) (stream.Blob, shared.BlobTrace, error) {
 		return nil, shared.NewBlobTrace(time.Since(start), d.Name()), errors.Err(err)
 	}
 	return blob, shared.NewBlobTrace(time.Since(start), d.Name()), nil
-}
-
-// Put stores the blob on disk
-func (d *DiskStore) Put(hash string, blob stream.Blob) error {
-	err := d.initOnce()
-	if err != nil {
-		return err
-	}
-
-	err = d.ensureDirExists(d.dir(hash))
-	if err != nil {
-		return err
-	}
-
-	// Open file with O_DIRECT
-	f, err := os.OpenFile(d.tmpPath(hash), openFileFlags, 0644)
-	if err != nil {
-		return errors.Err(err)
-	}
-	defer f.Close()
-
-	// Use directio writer
-	dio, err := directio.New(f)
-	if err != nil {
-		return errors.Err(err)
-	}
-	defer dio.Flush()
-	// Write the body to file
-	_, err = io.Copy(dio, bytes.NewReader(blob))
-	if err != nil {
-		return errors.Err(err)
-	}
-	err = os.Rename(d.tmpPath(hash), d.path(hash))
-	return errors.Err(err)
 }
 
 // PutSD stores the sd blob on the disk
