--- store/flags_darwin.go
@@ -1,9 +1,0 @@
-// +build darwin
-
-package store
-
-import (
-	"os"
-)
-
-var openFileFlags = os.O_WRONLY | os.O_CREATE
